+++ client/message/index.ts
@@ -1,3 +1,3 @@
 setTimeout(() => {
     document.getElementById("redirect")?.click();
-}, 1000 * 3);
+}, 1000);