+++ server/constants.ts
@@ -1,3 +1,3 @@
 export default {
-    VERSION: "0.5.0"
+    VERSION: "0.5.1"
 }