+++ server/routes/create.ts
@@ -4,6 +4,7 @@ import path from "path";
 import {customAlphabet} from "nanoid";
 import fileUpload, {UploadedFile} from "express-fileupload";
 import {PermissionCodes} from "../permissions";
+import bcrypt from "bcrypt";
 
 const DATA_DIR = "fdata";
 const FILE_DIR = "files";
@@ -40,7 +41,7 @@ export default (ctx:Context) => {
             res.render("message", {message: "You don't have permission to post!", redirectto: "/"});
         }
 
-        if(!res.locals.captchavalid) return res.render("/post", {msg: res.locals.captchamsg});
+        if(!res.locals.captchavalid) return res.render("message", {message: res.locals.captchamsg, redirectto: "/post"});
     
         const {title, body} = req.body;
     
@@ -75,7 +76,7 @@ export default (ctx:Context) => {
     app.post("/profile", async (req, res) => {
         if(!res.locals.loggedin) return res.redirect("/login");
     
-        const {bio, email} = req.body;
+        const {bio, email, password, newpassword, confnewpassword} = req.body;
     
         if(bio) {
             await dbc.updateById("users", res.locals.id, {about:bio});
@@ -84,7 +85,6 @@ export default (ctx:Context) => {
             let parts = email.split("@");
             if(parts[1] && parts[0] > 0 && parts[0] < 65 && parts[1] > 0 && parts[1] < 256) await dbc.updateById("users", res.locals.id, {email});
         }
-    
         if(req.files) if(req.files.file) {
             let mime = (req.files.file as UploadedFile).mimetype
             if(!(mime == "image/png" || mime == "image/jpeg" || mime == "image/gif" || mime == "image/bmp" || mime == "image/webp")) return res.redirect("/profile");
@@ -92,8 +92,20 @@ export default (ctx:Context) => {
             let file = await manipFile(req, ctx);
             await dbc.updateById("users", res.locals.id, {avatar:file});
         }
+        if(password && newpassword && confnewpassword) {
+            const pword = await dbc.getById("passwords", res.locals.id);
+            let result = bcrypt.compareSync(password, pword.hash);
+            if(!result) return res.render("message", {message: "Your password is incorrect!", redirectto: "/profile"});
+
+            if(newpassword !== confnewpassword) return res.render("message", {message: "Passwords do not match!", redirectto: "/profile"});
+            let salt = bcrypt.genSaltSync(10);
+            let hash = bcrypt.hashSync(newpassword, salt);
+            await dbc.updateById("passwords", res.locals.id, {hash});
+
+            return res.render("message", {message: "Successfully changed password", redirectto: `/users/${res.locals.id}`});
+        }
     
-        res.redirect(`/users/${res.locals.id}`);
+        return res.redirect(`/users/${res.locals.id}`);
     });
     
     app.get("/comment", (req, res) => {
@@ -110,7 +122,7 @@ export default (ctx:Context) => {
             return res.render("message", {message: "You don't have permission to comment!", redirectto: `/posts/${id}`});
         }
 
-        if(!res.locals.captchavalid) return res.redirect(`/posts/${id}/`);
+        if(!res.locals.captchavalid) return res.render("message", {message: res.locals.captchamsg, redirectto: `/posts/${id}/`});
     
         let post = await dbc.getById("posts", id);
         if(!post) return res.redirect("/");