+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 0.5.1
+- Added capability to change password (in profile)
+- Moving away from using register/login message (now redirects to message page)
+
 # 0.5.0
 - Started versioning system, added a database migration system for future database changes
 - Forum so far features: users, posts, comments, attachments, and permissions.